+++ src/levianeer/draconis/data/campaign/ids/Factions.java
@@ -4,5 +4,4 @@ public class Factions {
 
     public static final String DRACONIS = "XLII_draconis";
     public static final String FORTYSECOND = "XLII_fortysecond";
-    public static final String FIRSTFLEET = "XLII_firstfleet";
 }